Refactor code to utilize C# 10 file-scoped namespaces
This commit updates the repository to use the file-scoped namespaces introduced in C# 10. All classes were altered to use these simpler, less indented namespaces which makes the codebase cleaner and easier to read overall. The changes affect all the classes, controllers and the DbContext in the EF Core Stored Procedure tutorial project.